+++ examples/tremor/src/App.tsx
@@ -1,0 +1,34 @@
+import { QueryBuilderTremor } from '@react-querybuilder/tremor';
+import { useState } from 'react';
+import type { Field, RuleGroupType } from 'react-querybuilder';
+import { QueryBuilder, formatQuery } from 'react-querybuilder';
+import './styles.scss';
+
+const fields: Field[] = [
+  { name: 'firstName', label: 'First Name' },
+  { name: 'lastName', label: 'Last Name' },
+];
+
+const initialQuery: RuleGroupType = {
+  combinator: 'and',
+  rules: [
+    { field: 'firstName', operator: 'beginsWith', value: 'Stev' },
+    { field: 'lastName', operator: 'in', value: 'Vai,Vaughan' },
+  ],
+};
+
+export const App = () => {
+  const [query, setQuery] = useState(initialQuery);
+
+  return (
+    <div>
+      <QueryBuilderTremor>
+        <QueryBuilder fields={fields} query={query} onQueryChange={setQuery} />
+      </QueryBuilderTremor>
+      <h4>Query</h4>
+      <pre>
+        <code>{formatQuery(query, 'json')}</code>
+      </pre>
+    </div>
+  );
+};