+++ packages/react-querybuilder/genericTests/testValueSelector.tsx
@@ -13,6 +13,7 @@ import { findSelect, hasOrInheritsClass, userEventSetup } from './utils';
 
 type ValueSelectorTestsToSkip = Partial<{
   multi: boolean;
+  optgroup: boolean;
 }>;
 
 export const defaultValueSelectorProps = {
@@ -50,18 +51,20 @@ export const testSelect = (
       expect(getSelect().querySelectorAll('option')).toHaveLength(testValues.length);
     });
 
-    it('should render optgroups', () => {
-      const optGroups = [
-        { label: 'Test Option Group', options: 'values' in props ? props.values : props.options },
-      ];
-      const newProps =
-        'values' in props ? { ...props, values: optGroups } : { ...props, options: optGroups };
-      render(<Component {...newProps} />);
-      const getSelect = () => findSelect(screen.getByTitle(title));
-      expect(getSelect).not.toThrow();
-      expect(getSelect().querySelectorAll('optgroup')).toHaveLength(optGroups.length);
-      expect(getSelect().querySelectorAll('option')).toHaveLength(testValues.length);
-    });
+    if (!skip.optgroup) {
+      it('should render optgroups', () => {
+        const optGroups = [
+          { label: 'Test Option Group', options: 'values' in props ? props.values : props.options },
+        ];
+        const newProps =
+          'values' in props ? { ...props, values: optGroups } : { ...props, options: optGroups };
+        render(<Component {...newProps} />);
+        const getSelect = () => findSelect(screen.getByTitle(title));
+        expect(getSelect).not.toThrow();
+        expect(getSelect().querySelectorAll('optgroup')).toHaveLength(optGroups.length);
+        expect(getSelect().querySelectorAll('option')).toHaveLength(testValues.length);
+      });
+    }
 
     // Test as multiselect for <ValueEditor type="multiselect" /> and <ValueSelector />
     if (